Moving AXE, A2GPxQ, etc. to brevitas_examples
